+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Internalize UserAttributeService and remove getUserAttribute method. Note that the service is only intended to 
+* Remove UserAttributeService::getUserAttribute method. Note that the service is only intended to 
 deliver the currently logged-in user's attributes
 
 ## v0.1.198